Update model module - add code for Fit function, rename vars, etc.
- Rename many of the internal variables used in the model class to
  denote that they are private members

- Add code to fit the model into the code

*** Right now, I'm not completely happy with the API- part of me feels
like the data should be passed as part of the Fit () function, rather
than in the constructor.  The original consideration was that since
the data is integral to the functioning of the class later on (i.e. it
is impossible to use the class for forecasting / backtesting /
etc. without the data (since it is necessary for constructing
sequences), it belongs in the constructor).  The transformations are
also critical to the class and forms part of its invariants.  It would
be strange, I think, to inject the data as part of a given function
call.  Particularly if the function does not need to be called---such
as when a model is restored from disk